# fix: use Buffer.concat on node as it is faster