--- a/src/compare.ts
+++ b/src/compare.ts
@@ -2,6 +2,10 @@
  * Can be used with Array.sort to sort and array with Uint8Array entries
  */
 export function compare (a: Uint8Array, b: Uint8Array): number {
+  if (globalThis.Buffer != null) {
+    return globalThis.Buffer.compare(a, b)
+  }
+
   for (let i = 0; i < a.byteLength; i++) {
     if (a[i] < b[i]) {
       return -1