--- a/test/concat.spec.ts
+++ b/test/concat.spec.ts
@@ -1,6 +1,7 @@
 /* eslint-env mocha */
 
 import { expect } from 'aegir/chai'
+import { alloc } from '../src/alloc.js'
 import { concat } from '../src/concat.js'
 
 describe('Uint8Array concat', () => {
@@ -20,25 +21,9 @@ describe('Uint8Array concat', () => {
     expect(concat([a, b], 8)).to.deep.equal(c)
   })
 
-  it('concats mixed Uint8Arrays and Arrays', () => {
-    const a = Uint8Array.from([0, 1, 2, 3])
-    const b = [4, 5, 6, 7]
-    const c = Uint8Array.from([0, 1, 2, 3, 4, 5, 6, 7])
-
-    expect(concat([a, b])).to.deep.equal(c)
-  })
-
-  it('concats mixed Uint8Arrays and Arrays with a length', () => {
-    const a = Uint8Array.from([0, 1, 2, 3])
-    const b = [4, 5, 6, 7]
-    const c = Uint8Array.from([0, 1, 2, 3, 4, 5, 6, 7])
-
-    expect(concat([a, b], 8)).to.deep.equal(c)
-  })
-
   it('concat returns Uint8Array', () => {
     const a = Uint8Array.from([0, 1, 2, 3])
-    const b = [4, 5, 6, 7]
+    const b = alloc(10).fill(1)
     const c = concat([a, b])
     const slice = c.slice()